% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtracker.R
\name{assignWells}
\alias{assignWells}
\title{Takes xy data and assigns to wells for 24 or 48 well plate.}
\usage{
assignWells(files = NULL, wells = 24, markededges = TRUE,
  filenamediscard = ".txt", exportdata = TRUE, exportplot = TRUE,
  quiet = FALSE)
}
\arguments{
\item{files}{A character or vector of paths or filenames. See details.}

\item{wells}{A numeric indicating plate format. Either 24 or 48.}

\item{markededges}{A logical indicating if corners of plates have been marked. See details.}

\item{filenamediscard}{A character for part of the filename to be removed. '.txt' is removed by default.}

\item{exportdata}{A logical indicating if data table must be exported as a text file to the working directory.}

\item{exportplot}{A logical if results must be plotted as a figure and exported to working directory.}

\item{quiet}{A logical indicating if messages should be printed to console during the run. If \code{FALSE}, all output to console is killed except progress bar.}
}
\value{
Returns a dataframe with original data (x,y,slice) along with a new columns row, plate and well.
plate column is the filename without filenamediscard. The well column contains the well assignment.
If \code{exportdata=TRUE}, then a text file with the dataframe is exported in the working directory for each input file.
}
\description{
Takes xy data and assigns to wells for 24 or 48 well plate.
}
\details{
\strong{files}\cr
A character or vector of paths to files. On windows, use \code{choose.files()} for interactive selection.
The input files must be tab-delimited decimal as dot (.) text files. The file
must contain a minimum of 3 columns named x, y and slice. x is a numeric
indicating x coordinate and y is a numeric indicating y coordinate of each spot.
slice indicates the frame number for each spot. Any extra columns are not used.\cr
\strong{markededges}\cr
\code{markededges=T} indicates that the four corners of the plate have been marked in frame 1. They
will be used for plate alignment and plotting and will be removed from other analyses.\cr
}

