% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtracker.R
\name{gridWells}
\alias{gridWells}
\title{Internal: Computes plate layout and wells from marked data.}
\usage{
gridWells(dframe = NULL, wells = NULL)
}
\arguments{
\item{dframe}{A dataframe with x and y coordinates.}

\item{wells}{An integer indicating number of wells on the plate.}
}
\value{
Returns a dataframe with columns:
well,horizontal position 1 (h1), horizontal position 2 (h2), vertical position 1 (v1),
vertical position 2 (v2), well midpoint x (x), well midpoint y (y), top left label position x (labx),
top left label position y (laby) and bottom label y (laby1).
}
\description{
Internal: Computes plate layout and wells from marked data.
}
\details{
The 4 marked edges of the plate is used to compute well layout positions. The resulting output is used in plotting and other functions.
}

