% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtracker.R
\name{linkFrames}
\alias{linkFrames}
\title{Takes xy data with wells and links spots frame to frame.}
\usage{
linkFrames(files = NULL, dframe = NULL, wells = 24,
  filenamediscard = "-AssignedWells.txt", exportdata = TRUE,
  exportplot = TRUE, follow = "none", quiet = FALSE)
}
\arguments{
\item{files}{A character or vector of paths or filenames. A text file exported from \code{assignWells()} is suitable. See details.}

\item{dframe}{A dataframe output from \code{assignWells()}. See details.}

\item{wells}{A numeric indicating plate format. Either 24 or 48.}

\item{filenamediscard}{A character for part of the filename to be removed. '-AssignedWells.txt' is removed by default.}

\item{exportdata}{A logical indicating if data table must be exported as a text file to the working directory.}

\item{exportplot}{A logical if results must be plotted as a figure and exported to working directory.}

\item{follow}{A character indicating if the algorithm must be followed at every step. See details.}

\item{quiet}{A logical indicating if messages should be printed to console during the run. If \code{FALSE}, all output to console is killed except progress bar.}
}
\value{
Returns a dataframe with original data (plate,x,y,well) along with a new columns id, linktype and frame.
plate column is the filename without filenamediscard. The well column contains the well assignment. If number
of input files > 1, then the dataframe RESULTS FOR THE LAST INPUT FILE is returned. If \code{exportdata=TRUE}, then a text file ending in '-Tracks' with
dataframe is exported in the working directory for each input file. This '-Tracks' text file is used as input for several other functions.
}
\description{
Takes xy data with wells and links spots frame to frame.
}
\details{
Note that the 'xx-EdgeMarks.txt' file must be present in the working directory for the plot exports to work.
\strong{Input}\cr
The input for this function can be a vector of filenames passed to argument \code{files} or
a dataframe passed to argument \code{dframe}. Use any one and not both.
\strong{files}\cr
A character or vector of paths to files. On windows, use \code{choose.files()} for interactive selection.
The input files must be tab-delimited decimal as dot (.) text files. The file
must contain a minimum of 4 columns named plate, x, y and well. plate is a character
denoting the filename. x is a numeric indicating x coordinate and y is a numeric
indicating y coordinate of each spot. well indicates the well number to which each spot
is assigned. Any extra columns are not used.\cr
\strong{dframe}\cr
A dataframe containing a minimum of 4 columns named plate, x, y and well. plate is a character
denoting the filename. x is a numeric indicating x coordinate and y is a numeric
indicating y coordinate of each spot. well indicates the well number to which each spot
is assigned. Any extra columns are not used.\cr
\strong{follow}\cr
Set as 'none', 'interactive' or 'track'.
In 'interactive' mode, a plot is shown at every frame and waits for user input.
In 'track' mode, the track path creation is shown in real-time.
}

