% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtracker.R
\name{spotRatio}
\alias{spotRatio}
\title{Computes larval presence in center vs edges of the wells.}
\usage{
spotRatio(files = NA, wells = 24, filenamediscard = "-Tracks.txt",
  mm = 5.4, outercircleradius = 8.1, innercircleradius = 4.8,
  exportdata = TRUE, exportplot = TRUE, quiet = FALSE)
}
\arguments{
\item{files}{A character or vector of paths or filenames. See details.}

\item{wells}{A numeric indicating plate format. Either 24 or 48.}

\item{filenamediscard}{A character for part of the filename to be removed. '-Tracks.txt' is removed by default.}

\item{mm}{A numeric indicating number of pixels in 1 mm.}

\item{outercircleradius}{A numeric indicating radius of outer circle (radius of one well) in mm.}

\item{innercircleradius}{A numeric indicating radius of inner circle in mm.}

\item{exportdata}{A logical indicating if data table must be exported as a text file to the working directory.}

\item{exportplot}{A logical if results must be plotted as a figure and exported to working directory.}

\item{quiet}{A logical indicating if messages should be printed to console during the run. If \code{FALSE}, all output to console is killed except progress bar.}
}
\value{
Returns a dataframe with columns: row, plate names, wells, number of outer spots, inner spots and spots ratio.
If \code{exportdata=TRUE}, the dataframe is export as a text file in the working directory for each input file.
}
\description{
Counts spots in the center vs edges of the wells and computes a ratio.
}
\details{
Note that the 'xx-EdgeMarks.txt' file must be present in the working directory for the plot exports to work.\cr
\strong{files}\cr
A character or vector of paths to files. On windows, use \code{choose.files()} for interactive selection.
The input files must be tab-delimited decimal as dot (.) text files. The file
must contain a minimum of 3 columns named x, y and well. x is a numeric
indicating x coordinate and y is a numeric indicating y coordinate of each spot. well indicates well assignment for each spot.
Any extra columns are not used. Use output from \code{linkFrames()}.\cr
A text file with edge marks (filename-EdgeMarks.txt) must be present in the working directory for plotting. The edge marks file is exported by function \code{assignWells()}.
}

