% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drtracker.R
\name{trackFeatures}
\alias{trackFeatures}
\title{Computes features for tracks such as distance, speed etc.}
\usage{
trackFeatures(files = NULL, wells = 24, fps = 25, mm = 5.4,
  activitydist = 5, filenamediscard = "-Tracks.txt", exportdata = TRUE,
  exportplot = TRUE, quiet = FALSE)
}
\arguments{
\item{files}{A character or vector of paths or filenames. A text file exported from \code{linkFrames()} is suitable. See details.}

\item{wells}{A numeric indicating plate format. Either 24 or 48.}

\item{fps}{A numeric indicating framerate of video in number of frames per second.}

\item{mm}{A numeric indicating number of pixels in 1 mm.}

\item{activitydist}{A numeric indicating distance in mm after which the larvae is considered active.}

\item{filenamediscard}{A character for part of the filename to be removed. '-Tracks.txt' is removed by default.}

\item{exportdata}{A logical indicating if data table must be exported as a text file to the working directory.}

\item{exportplot}{A logical if results must be plotted as a figure and exported to working directory.}

\item{quiet}{A logical indicating if messages should be printed to console during the run. If \code{FALSE}, all output to console is killed except progress bar.}
}
\value{
Returns a dataframe with columns: plate name, id, well, total distance in pixels and mm, mean speed in pixels per sec (pps) and
in mm per second (mmps), max speed in pixels per sec (pps) and mm per sec (mmps), activity (active seconds), duration of
the whole sequence in frames (fr) and seconds (sec), framerate (fps) and calibration,
number of pixels in one mm (mm).
}
\description{
Computes features for tracks such as distance, speed etc.
}
\details{
\strong{files}\cr
A character or vector of paths to files. On windows, use \code{choose.files()} for interactive selection.
The input files must be tab-delimited decimal as dot (.) text files. The file
must contain a minimum of 5 columns named plate, x, y, well and frame. plate is a character
denoting the filename. x is a numeric indicating x coordinate and y is a numeric
indicating y coordinate of each spot. well indicates the well number to which each spot
is assigned. frame indicates frame number for each spot. Any extra columns are not used.\cr
The number of wells are determined from input data.
}

