% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drtracker.R
\name{ltrack}
\alias{ltrack}
\title{Track larval movement in 24 or 48 well assay plates.}
\usage{
ltrack(files = NULL, wells = 24, markededges = TRUE, fps = 25,
  mm = 5.4, coverage = TRUE, msd = TRUE, alphahull = TRUE,
  alphavalue = 4, filenamediscard = ".txt", exportdata = TRUE,
  exportplot = TRUE, follow = "none", centerwell = FALSE)
}
\arguments{
\item{files}{A character or vector of paths or filenames. See details.}

\item{wells}{A numeric indicating plate format. Either 24 or 48.}

\item{markededges}{A logical indicating if corners of plates have been marked. See details.}

\item{fps}{A numeric indicating framerate of video in number of frames per second.}

\item{mm}{A numeric indicating number of pixels in 1 mm.}

\item{coverage}{A logical indicating if the coverge must be computed. See details.}

\item{msd}{A logical indicating if the minimum spanning distance should be computed. See details.}

\item{alphahull}{A logical indicating if alphahull of point cloud is to be calculated. See details.}

\item{alphavalue}{A numeric indicating alpha value for alphahull. See details.}

\item{filenamediscard}{A character for part of the filename to be removed.}

\item{exportdata}{A logical indicating if data tables must be exported to working directory.}

\item{exportplot}{A logical if results must be plotted and exported to working directory.}

\item{follow}{A character indicating if the algorithm must be followed at every step.
Options are 'interactive', 'track' or 'none'. See details.}

\item{centerwell}{A logical indicating if the well number must be printed at center of the well on plots.}
}
\value{
Returns a list containing two components; 'RawData' and 'Track statistics'.\cr
The 'RawData' is a dataframe for one input file or a list for 2 or more input files.
The 'RawData' contains xy coordinates, slice, well, id and comment per spot. The
id links the spots as tracks. The comment shows the method used to link a spot.
'Single' means that there was only one spot in the next frame. This may be the correct
spot or a stray spot. 'Nearest' means there was more than one spot in the next frame
and the nearest spot was chosen. 'Duplicate' means that no spot was present in the next frame
and the previous spot was duplicated in next frame.\cr
'Track Statistics' contains plate name, id, well, total distance in pixels and mm, mean speed in pixels per sec (pps) and
in mm per second (mmps), max speed in pixels per sec (pps) and mm per sec (mmps), duration of
the whole sequence in frames (fr) and seconds (sec), framerate (fps) and calibration,
number of pixels in one mm (mm). If \code{coverage=TRUE}, then coverage_pxsq is added.
If \code{msd=TRUE}, then msd_px and msd_mm are added. If \code{alphahull=TRUE}, then alphahull_pxsq is added.\cr
If \code{exportdata = T}, then three text files are exported: Wells, Tracks and TrackStatistics.
If more than one file was selected, a Combined-TrackStatistics file is also exported.
If \code{exportplot = T}, then three-five figures are exported: EdgeSpots (if edge spots are used),
Wells & Spots, Wells & Tracks, Coverage and Track Statistics.
}
\description{
Tracks single larvae in 24 or 48 well assay plates from xy spot data
and computes distance and speed. Exports data as text files and generates plots.
}
\details{
The quality of tracks almost completely depends on the image thresholding and subsequent xy data.\cr
\strong{files}\cr
The input files must be tab-delimited decimal as dot (.) text files. The file
must contain a minimum of 3 columns named x, y and slice. x is a numeric
indicating x coordinate and y is a numeric indicating y coordinate of each spot.
slice indicates the frame number for each spot. Extra columns are not used.\cr
\strong{markededges}\cr
\code{markededges = T} indicates that the four corners of the plate have been marked in frame 1. They
will be used for plate alignment and plotting and will be removed from analyses.\cr
\strong{coverage}\cr
Area covered by each larvae in their respective well across the whole duration.
Computed from convex hull of points. The polygon area is computed based on the
function \code{polyarea} from package \code{pracma}.\cr
\strong{msd}\cr
The minimum spanning distance based on the minimum spanning network of point cloud.
COmputed using \code{spantree} from package \code{vegan}.
\strong{alphahull}\cr
The alphahull of point cloud based on the alphavalue. The function \code{ahull} from
package \code{alphahull}.
\strong{filenamediscard}\cr
The file name of the input file is used on plots and text files for identification.
Part of the filename to be removed can be indicated here. '.txt' is removed by default.\cr
\strong{follow}\cr
In 'interactive' mode, a plot is shown at every frame and waits for user input.
In 'track' mode, the track path creation is shown in real-time.
}

